+++ lib/server/Recalc/Server/Types.hs
@@ -1,0 +1,34 @@
+{-# OPTIONS_GHC -fno-warn-orphans #-}
+
+module Recalc.Server.Types where
+
+import Data.Aeson qualified as Json
+
+import Recalc.Engine (Isn't (..))
+
+instance Isn't Json.Value where
+  isn't = (Json.Null ==)
+
+data Nullable t
+  = Missing
+  | Null
+  | Is t
+  deriving (Eq, Show)
+
+instance Isn't a => Isn't (Nullable a) where
+  isn't (Is x) = isn't x
+  isn't _ = True
+
+instance Json.FromJSON t => Json.FromJSON (Nullable t) where
+  parseJSON Json.Null = pure Null
+  parseJSON v = Is <$> Json.parseJSON v
+
+  omittedField = Just Missing
+
+instance Json.ToJSON t => Json.ToJSON (Nullable t) where
+  toJSON = \case
+    Null -> Json.Null
+    Missing -> Json.Null
+    Is t -> Json.toJSON t
+
+  omitField = \case Missing -> True; _ -> False