+++ recalc-vscode/src/frontend/plugin.ts
@@ -1,7 +1,5 @@
 import { Dependency, Inject, Injector, Plugin, UniverInstanceType } from '@univerjs/core';
 import { ComponentManager } from '@univerjs/ui';
-
-import { RefSelectionsRenderService } from '@univerjs/sheets-formula-ui';
 import { IRenderManagerService } from '@univerjs/engine-render';
 
 import { MessageController } from './controllers/rpc.controller';
@@ -19,23 +17,10 @@ export class RecalcPlugin extends Plugin {
   }
 
   override onStarting() {
-    ([
-      [MessageController],
-    ] as Dependency[]).forEach(dep => this._injector.add(dep))
+    ([ [MessageController] ] as Dependency[]).forEach(dep => this._injector.add(dep))
   }
 
   override onReady() {
     this._injector.get(MessageController);
-  }
-
-  override onRendered(): void {
-    ([
-      [RefSelectionsRenderService],
-     ] as Dependency[]).forEach((dep) =>
-      // register and make sure things are disposed in time
-      this.disposeWithMe(
-        this._renderManagerService.registerRenderModule(UniverInstanceType.UNIVER_SHEET, dep)
-      )
-    );
   }
 }